# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Uploy API
  version: 0.2.0
  description: Uploy deployment platform API
paths:
...
  /api/services/{id}/envs:
    get:
      operationId: listServiceEnvs
      summary: List environment variables for a service
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of environment variables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceEnvResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: upsertServiceEnv
      summary: Set or update an environment variable
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertEnvRequest"
      responses:
        "200":
          description: Environment variable set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceEnvResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/services/{id}/pending-changes:
    get:
      operationId: getServicePendingChanges
      summary: List the changes waiting to be deployed for a service
      description: >
        Itemises what the running container would have to change to match the
        service as it is configured now. Owner and developer only: the list
        carries environment variable values, which is the same reason
        listServiceEnvs is gated.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The pending changes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingChangesResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        code:
          type: string
          description: >
            Machine-readable failure stage. Present on SSH probe failures so a
            client can offer the right remedy: the three stages need entirely
            different fixes, and the human message alone cannot be branched on.
          enum:
            - unreachable
            - key_rejected
            - key_invalid
            - session_failed
            - docker_missing
...
    ConfigChange:
      type: object
      required:
        - key
        - label
        - type
      properties:
        key:
          type: string
          description: >
            Stable identifier for this change, unique within the list —
            "image", "domain:app.example.com", "env:DATABASE_URL".
        label:
          type: string
          description: What the change is called in the interface.
        type:
          type: string
          enum:
            - added
            - removed
            - changed
        old_value:
          type: string
          description: Absent when the thing did not exist before.
        new_value:
          type: string
          description: Absent when the thing no longer exists.
    PendingChangesResponse:
      type: object
      required:
        - has_baseline
        - changes
      properties:
        has_baseline:
          type: boolean
          description: >
            False when there is no deployed configuration to compare against:
            the service has never deployed, or last deployed before Uploy
            recorded configurations. The service is pending either way, but
            changes is empty because no itemised answer exists — say so rather
            than showing an empty list under a badge that says otherwise.
        changes:
          type: array
          items:
            $ref: "#/components/schemas/ConfigChange"
...
    UpsertEnvRequest:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
        value:
          type: string
    ServiceEnvResponse:
      type: object
      required:
        - id
        - key
        - value
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
        value:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time